Make Input component controlled (#406)
Hey,

I was playing with the "TodoMVC" example and tried to reset the input when a todo is added. I thought that resetting the `inputValue` state in the `AddTodo` action would make it. But I noticed that the `inputValue` state was not even passed to the `Input` component.

It seems that the `Input` component manages his own state. I changed that to make it a controlled component which makes more sense to me.

If I'm not mistaken, you have both in react:
* Uncontrolled if you don't pass a `value` prop
* Controlled if we pass a `value` prop

I don't know if it really makes sense to have both possibilities ? Even if what I did make sens for you? 

If yes, I've seen that `Checkbox` is also `uncontrolled` (and probably others too) and I would be happy to change them as well :)