Bugfix: Disable subpixel text until we have a configuration story (#752)
Skia supports subpixel character position rendering via setSubpixelText, but this can affect measurement and metrics.

On some platforms, subpixel rendering isn't supported - so this can skew our measurements when the rendered text doesn't align with the measurements we expect. For now, I'll turn it off, until we can reliably detect when it will be enabled, or have a configuration story for it.

Note that this doesn't disable anti-aliasing, or lcd striping - just subpixel positioning.