Spring: Add enabled property to spring; fix bug with rest (#764)
This change adds an enabled property to the spring.

In addition, I noticed a bug with the restThreshold - once the rest threshold has been reached, there would be no further animation. This is OK for low values of restThreshold, where the spring would've been close to its final destination, but for higher values, the spring may stop animating far away from the destination. This clamps the spring in the case there is no animation (ie, if it is resting, or not active).