feat(windows): ANGLE support (#983)
On Windows, Direct3D is much better supported than OpenGL - it's often the case where Direct3D is available, but OpenGL isn't (for example - on VMs). In addition, the driver support tends to be for Direct3D on Windows.

Google built a library called [ANGLE](https://github.com/google/angle) which provides an OpenGL ES API backed by Direct3D (It's used by Chrome as well to back WebGL).

This brings in ANGLE to be used as the renderer back-end for Revery and Skia. It requires bringing in `libEGL.dll` and `libGLESV2.dll`, and then setting up SDL to render to an OpenGL ES context.

Switching to ANGLE will likely fix several issues running Onivim 2 on Windows w/o up-to-date OpenGL drivers:
- https://github.com/onivim/oni2/issues/1522
- https://github.com/onivim/oni2/issues/1809
- https://github.com/onivim/oni2/issues/1842

(at least, enough to get the app visible & running - I can run Onivim 2 and the revery sample app on a Windows 10 VM w/ this change!)