feat(style): Bring back `opacity` style (#1002)
The `opacity` style option was removed in #540 , in trying to factor to semantic components. However, since we don't have that yet (dependent on some larger refactorings, as discussed in #489 ) - we should bring this back.

I'd prefer to use a more semantic component like `<Opacity />`, however, in the current design, that can impact layout, as every primitive has an associated layout node.