chore(ci): Move to single CI strategy (Azure Pipelines) (#1022)
This is work to unblock our PR builds. It looks like the first issue is the `setup-node` task being out of date, but then there are still additional failures in the docker images beyond that.

I'm going to revert the GHA for now, and just stick with Azure CI, since it's the same build infra we ship Onivim from - this will remove some maintenance burden from me.